Move asm definitions to the ty module.
