+++ sway-core/src/language/ty/expression/mod.rs
@@ -1,6 +1,8 @@
+mod asm;
 #[allow(clippy::module_inception)]
 mod expression;
 mod expression_variant;
 
+pub use asm::*;
 pub use expression::*;
 pub use expression_variant::*;
