+++ sway-core/src/semantic_analysis/ast_node/expression/typed_expression.rs
@@ -792,7 +792,7 @@ impl ty::TyExpression {
             .registers
             .into_iter()
             .map(
-                |AsmRegisterDeclaration { name, initializer }| TyAsmRegisterDeclaration {
+                |AsmRegisterDeclaration { name, initializer }| ty::TyAsmRegisterDeclaration {
                     name,
                     initializer: initializer.map(|initializer| {
                         let ctx = ctx
