+++ sway-core/src/semantic_analysis/ast_node/expression/typed_expression_variant.rs
@@ -38,31 +38,3 @@ pub struct TypeCheckedStorageAccessDescriptor {
     pub(crate) type_id: TypeId,
     pub(crate) span: Span,
 }
-
-#[derive(Clone, Debug)]
-pub struct TyAsmRegisterDeclaration {
-    pub(crate) initializer: Option<ty::TyExpression>,
-    pub(crate) name: Ident,
-}
-
-// NOTE: Hash and PartialEq must uphold the invariant:
-// k1 == k2 -> hash(k1) == hash(k2)
-// https://doc.rust-lang.org/std/collections/struct.HashMap.html
-impl PartialEq for TyAsmRegisterDeclaration {
-    fn eq(&self, other: &Self) -> bool {
-        self.name == other.name
-            && if let (Some(l), Some(r)) = (self.initializer.clone(), other.initializer.clone()) {
-                l == r
-            } else {
-                true
-            }
-    }
-}
-
-impl CopyTypes for TyAsmRegisterDeclaration {
-    fn copy_types(&mut self, type_mapping: &TypeMapping) {
-        if let Some(ref mut initializer) = self.initializer {
-            initializer.copy_types(type_mapping)
-        }
-    }
-}
